requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==2.0.5
django-cors-headers==2.2.0
django-extra-fields==1.0.0
django-filter==1.1.0
djangorestframework==3.8.2
drf-writable-nested==0.4.2
enum34==1.1.6
ipython-genutils==0.1.0
mysqlclient==1.3.12
nameparser==0.5.7
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
ply==3.11
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pyasn1==0.4.3
pycryptodome==3.6.0
pycryptodomex==3.6.3
Pygments==2.1.3
pysmi==0.3.1
pysnmp==5.0.0
pytz==2018.4
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
